# openAPI.yaml
openapi: 3.0.0
info:
  title: Ekadashi-server
  version: 0.0.1
paths:
  /register:
    post:
      summary: Register a new user in ekadashi database.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registration succeed
        '400':
          description: Incorrect username or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Incorrect username or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeed
          headers:
            Set-Cookie:
              schema:
                type: string
                example: session_token=abcde12345
        '400':
          description: Cannot decode user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Incorrect username or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /ekadashi/next:
    get:
      security:
        - cookieAuth: []
      summary: Show next ekadashi day
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
        '500':
          description: Internal server error
components:
  schemas:
    RegisterRequest:
      allOf:
        - $ref: '#/components/schemas/LoginRequest'
    LoginRequest:
      properties:
        username:
          type: string
          minLength: 6
          nullable: false
          pattern: '`^[a-zA-Z1-9]+$`'
        password:
          type: string
          minLength: 6
          nullable: false
          pattern: '`^[a-zA-Z1-9]+$`'
      required:
        - 'username'
        - 'password'
    ErrorResponse:
      properties:
        reason:
          type: string
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: 'session_token'